requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
transformers==4.35.0
torch==2.2.0+cpu
tensorflow==2.15.0
tf-keras==2.15.0
Pillow==10.1.0
pytesseract==0.3.10
python-docx==1.0.1
PyPDF2==3.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 